Fix bug when pushing to a pull request which enabled dismiss approval automatically
